test create account api endpoint :tada:
