--- banka/api/tests/mock_data/user_data.py
@@ -1,15 +1,0 @@
-
-
-class UserData:
-
-    @staticmethod
-    def user_signup_data():
-        return {
-           "firstname": "test",
-            "lastname": "testing",
-            "password": "testsecret",
-            "email": "test@gmail.com",
-            "user_type": "staff",
-            "is_admin": "true"
-        }
-
